+++ source/renderer/dialog/project-properties.js
@@ -35,7 +35,7 @@ class ProjectDialog extends ZettlrDialog {
     data.supportedPapertypes = SUPPORTED_PAPERTYPES
     data.papertypeNames = PAPERNAMES
     data.availableMarginUnits = ['cm', 'mm', 'pt']
-    data.availablePageNumberingSystems = ['arabic', 'alph', 'Alph', 'roman', 'Roman', 'gobble']
+    data.availablePageNumberingSystems = ['arabic', 'alph', 'alph_upper', 'roman', 'roman_upper', 'gobble']
     return data
   }